Conflicts:
	app/controllers/sessions_controller.rb
	app/helpers/application_helper.rb
	app/models/observation.rb
	app/views/admin/stats.html.erb
	app/views/calendars/compare.html.erb
	app/views/check_lists/_form.html.erb
	app/views/check_lists/show.html.erb
	app/views/comments/_comment.html.erb
	app/views/comments/_comment_form_fields.html.erb
	app/views/comments/_comments.html.erb
	app/views/devise/mailer/confirmation_instructions.html.erb
	app/views/emailer/invite.html.erb
	app/views/facebook/_albums.html.erb
	app/views/facebook/options.html.erb
	app/views/flags/_flag.html.erb
	app/views/flags/index.html.erb
	app/views/flags/new.html.erb
	app/views/flickr/_photos_form_field.html.erb
	app/views/flickr/create_invite.html.erb
	app/views/flickr/link.html.erb
	app/views/flickr/options.html.erb
	app/views/flickr/photos.html.erb
	app/views/flickr/success.html.erb
	app/views/help/getting_started.html.erb
	app/views/identifications/_identification.html.erb
	app/views/listed_taxa/_guide.html.erb
	app/views/listed_taxa/show.html.erb
	app/views/lists/_add_batch_dialog.html.erb
	app/views/lists/_compare_form.html.erb
	app/views/lists/_form_fields.html.erb
	app/views/lists/_listed_taxon.html.erb
	app/views/lists/add_taxon_batch.html.erb
	app/views/lists/by_login.html.erb
	app/views/lists/compare.html.erb
	app/views/observation_fields/edit.html.erb
	app/views/observation_fields/index.html.erb
	app/views/observation_fields/new.html.erb
	app/views/observation_fields/show.html.erb
	app/views/observations/_batch_form_row.html.erb
	app/views/observations/_citation.html.erb
	app/views/observations/_filter_fields.html.erb
	app/views/observations/_form_fields.html.erb
	app/views/observations/_location.html.erb
	app/views/observations/_new_nav.html.erb
	app/views/observations/_observation_component.html.erb
	app/views/observations/_observation_csv_parse_errors.html.erb
	app/views/observations/_observation_form_field_rows.html.erb
	app/views/observations/_observations_table_header.html.erb
	app/views/observations/_widget_code.html.erb
	app/views/observations/add_from_list.html.erb
	app/views/observations/add_nearby.mobile.erb
	app/views/observations/by_login.html.erb
	app/views/observations/edit.html.erb
	app/views/observations/identotron.html.erb
	app/views/observations/index.js.rjs
	app/views/observations/new.html.erb
	app/views/observations/show.html.erb
	app/views/photos/_photo.html.erb
	app/views/photos/_photo_list_form.html.erb
	app/views/photos/show.html.erb
	app/views/places/_create_external_place_links.html.erb
	app/views/places/_identotron.html.erb
	app/views/places/_search_form.html.erb
	app/views/places/new.html.erb
	app/views/places/show.html.erb
	app/views/posts/_form.html.erb
	app/views/posts/_post.html.erb
	app/views/posts/archives.html.erb
	app/views/posts/browse.html.erb
	app/views/posts/edit.html.erb
	app/views/posts/index.html.erb
	app/views/posts/new.html.erb
	app/views/posts/show.html.erb
	app/views/projects/_form.html.erb
	app/views/projects/browse.html.erb
	app/views/projects/join.html.erb
	app/views/projects/search.html.erb
	app/views/projects/show.html.erb
	app/views/projects/terms.html.erb
	app/views/quality_metrics/_quality_metric_row.html.erb
	app/views/quality_metrics/_quality_metrics.html.erb
	app/views/shared/_header.html.erb
	app/views/shared/_header.mobile.erb
	app/views/sources/_nested_form_fields.html.erb
	app/views/sources/index.html.erb
	app/views/taxa/_add_to_place_link.html.erb
	app/views/taxa/_browse.html.erb
	app/views/taxa/_description.html.erb
	app/views/taxa/_form.html.erb
	app/views/taxa/_merge.html.erb
	app/views/taxa/_photos.html.erb
	app/views/taxa/_search_form.html.erb
	app/views/taxa/_taxon.html.erb
	app/views/taxa/add_places.html.erb
	app/views/taxa/curation.html.erb
	app/views/taxa/edit.html.erb
	app/views/taxa/edit_photos.html.erb
	app/views/taxa/index.html.erb
	app/views/taxa/map.html.erb
	app/views/taxa/merge.html.erb
	app/views/taxa/search.html.erb
	app/views/taxa/show.html.erb
	app/views/taxa/show.mobile.erb
	app/views/taxa/tree.html.erb
	app/views/taxon_names/edit.html.erb
	app/views/taxon_ranges/_citation.html.erb
	app/views/user_mailer/activation.text.plain.erb
	app/views/user_mailer/admin_password_change.text.plain.erb
	app/views/user_mailer/signup_notification.text.plain.erb
	app/views/users/_new_form.html.erb
	app/views/users/curation.html.erb
	app/views/users/edit.html.erb
	app/views/users/edit_after_auth.html.erb
	app/views/users/index.html.erb
	app/views/users/new.html.erb
	app/views/users/sessions/new.html.erb
	app/views/users/show.html.erb

+++ app/models/photo.rb
@@ -1,4 +1,9 @@
+#encoding: utf-8
 class Photo < ActiveRecord::Base
+  # comment that's posted to invite someone to upload that photo to inat
+  # {{INVITE_LINK}} gets replaced with the url
+  DEFAULT_INVITE_COMMENT = "Great photo!  You should post it to iNaturalist.org: {{INVITE_LINK}}"
+
   belongs_to :user
   has_many :observation_photos, :dependent => :destroy
   has_many :taxon_photos, :dependent => :destroy
@@ -15,7 +20,7 @@ class Photo < ActiveRecord::Base
   cattr_accessor :descendent_classes
   cattr_accessor :remote_descendent_classes
   
-  before_save :set_license
+  before_save :set_license, :trim_fields
   after_save :update_default_license,
              :update_all_licenses
   
@@ -30,33 +35,31 @@ class Photo < ActiveRecord::Base
     4 => {:code => Observation::CC_BY,        :short => "CC BY",        :name => "Attribution License", :url => "http://creativecommons.org/licenses/by/3.0/"},
     5 => {:code => Observation::CC_BY_SA,     :short => "CC BY-SA",     :name => "Attribution-ShareAlike License", :url => "http://creativecommons.org/licenses/by-sa/3.0/"},
     6 => {:code => Observation::CC_BY_ND,     :short => "CC BY-ND",     :name => "Attribution-NoDerivs License", :url => "http://creativecommons.org/licenses/by-nd/3.0/"},
-    7 => {:code => "PD",                      :short => "PD",           :name => "Public domain, no known copyright restrictions", :url => "http://flickr.com/commons/usage/"}
+    7 => {:code => "PD",                      :short => "PD",           :name => "Public domain", :url => "http://en.wikipedia.org/wiki/Public_domain"},
+    8 => {:code => "GFDL",                    :short => "GFDL",         :name => "GNU Free Documentation License", :url => "http://www.gnu.org/copyleft/fdl.html"}
   }
   LICENSE_NUMBERS = LICENSE_INFO.keys
+  LICENSE_INFO.each do |code, info|
+    const_set info[:code].upcase.gsub(/\-/, '_'), code
+  end
+
+  SQUARE = 75
+  THUMB = 100
+  SMALL = 240
+  MEDIUM = 500
+  LARGE = 1024
+
+  validate :licensed_if_no_user
   
   def to_s
     "<#{self.class} id: #{id}, user_id: #{user_id}>"
   end
   
-  def validate
-    if user.blank? && self.license == 0
+  def licensed_if_no_user
+    if user.blank? && (license == COPYRIGHT || license.blank?)
       errors.add(
         :license, 
-        "must be a Creative Commons license if the photo wasn't added by " +
-        "an iNaturalist user using their linked Flickr account.")
-    end
-    
-    # Check to make sure the user owns the flickr photo
-    if self.user && self.api_response
-      if self.api_response.is_a?(Net::Flickr::Photo)
-        fp_flickr_user_id = self.api_response.owner
-      else
-        fp_flickr_user_id = self.api_response.owner.nsid
-      end
-      
-      unless fp_flickr_user_id == self.user.flickr_identity.flickr_user_id
-        errors.add(:user, "must own the photo on Flickr.")
-      end
+        "must be set if the photo wasn't added by an iNat user.")
     end
   end
   
@@ -64,6 +67,13 @@ class Photo < ActiveRecord::Base
     return true unless license.blank?
     return true unless user
     self.license = Photo.license_number_for_code(user.preferred_photo_license)
+    true
+  end
+
+  def trim_fields
+    %w(native_realname native_username).each do |c|
+      self.send("#{c}=", read_attribute(c).to_s[0..254]) if read_attribute(c)
+    end
     true
   end
   
@@ -76,9 +86,15 @@ class Photo < ActiveRecord::Base
     elsif (o = observations.first)
       o.user.name || o.user.login
     else
-      "anonymous Flickr user"
+      "anonymous"
     end
-    "#{license_short} #{name}"
+    if license == PD
+      "#{name}, no known copyright restrictions (#{license_name})"
+    elsif open_licensed?
+      "(c) #{name}, some rights reserved (#{license_short})"
+    else
+      "(c) #{name}, all rights reserved"
+    end
   end
   
   def license_short
@@ -104,23 +120,20 @@ class Photo < ActiveRecord::Base
   def creative_commons?
     license.to_i > COPYRIGHT && license.to_i < NO_COPYRIGHT
   end
+
+  def open_licensed?
+    license.to_i > COPYRIGHT && license != PD
+  end
   
   # Try to choose a single taxon for this photo.  Only works if class has 
   # implemented to_taxa
   def to_taxon
-    photo_taxa_from_scinames = try(:to_taxa, :lexicon => TaxonName::SCIENTIFIC_NAMES)
-    unless photo_taxa_from_scinames.blank?
-      unless photo_taxa_from_scinames.detect{|t| t.rank_level.blank?}
-        photo_taxa_from_scinames = photo_taxa_from_scinames.sort_by(&:rank_level)
-      end
-      return photo_taxa_from_scinames.detect(&:species_or_lower?) || photo_taxa_from_scinames.first
-    end
-    
-    photo_taxa = try(:to_taxa)
-    return nil if photo_taxa.blank?
-    unless photo_taxa.detect{|t| t.rank_level.blank?}
-      photo_taxa = photo_taxa.sort_by(&:rank_level)
-    end
+    return unless respond_to?(:to_taxa)
+    photo_taxa = to_taxa(:lexicon => TaxonName::SCIENTIFIC_NAMES, :valid => true, :active => true)
+    photo_taxa = to_taxa(:lexicon => TaxonName::SCIENTIFIC_NAMES) if photo_taxa.blank?
+    photo_taxa = to_taxa if photo_taxa.blank?
+    return if photo_taxa.blank?
+    photo_taxa = photo_taxa.sort_by{|t| t.rank_level || Taxon::ROOT_LEVEL + 1}
     photo_taxa.detect(&:species_or_lower?) || photo_taxa.first
   end
   
@@ -153,6 +166,25 @@ class Photo < ActiveRecord::Base
     return false if user.blank?
     user.id == user_id || observations.exists?(:user_id => user.id)
   end
+
+  def orphaned?
+    observation_photos_exist = observation_photos.loaded? ? observation_photos.size > 0 : observation_photos.exists?
+    taxon_photos_exist = taxon_photos.loaded? ? taxon_photos.size > 0 : taxon_photos.exists?
+    !observation_photos_exist && !taxon_photos_exist
+  end
+
+  def source_title
+    self.class.to_s.gsub(/Photo$/, '').underscore.humanize.titleize
+  end
+
+  def best_url(size = "medium")
+    size = size.to_s
+    sizes = %w(original large medium small thumb square)
+    size = "medium" unless sizes.include?(size)
+    size_index = sizes.index(size)
+    methods = sizes[size_index.to_i..-1].map{|s| "#{s}_url"} + ['original']
+    try_methods(*methods)
+  end
   
   # Retrieve info about a photo from its native source given its native id.  
   # Should be implemented by descendents
@@ -171,9 +203,7 @@ class Photo < ActiveRecord::Base
     photos = Photo.all(:conditions => ["id IN (?)", [ids].flatten])
     return if photos.blank?
     photos.each do |photo|
-      if !photo.observation_photos.exists? && !photo.taxon_photos.exists?
-        photo.destroy
-      end
+      photo.destroy if photo.orphaned?
     end
   end
