+++ tools/eol_collections_to_projects.rb
@@ -8,12 +8,10 @@ This script is meant to keep the EOL projects on iNaturalist in sync with the
 EOL collections in an EOL collection (yes, EOL has collections of collections)
 
 Usage:
-  script/runner tools/eol_collections_to_projects.rb <EOL collection url>
+  rails runner tools/eol_collections_to_projects.rb <EOL collection url>
 
 The script:
-   1) makes projects for collections if they don't already exist but doesn't
-     automatically add the icon, descripton, terms, or custom header - that
-     would be nice
+   1) makes projects for collections if they don't already exist
    2) adds listed_taxa to the list for the project for taxa in the EOL 
       collection
    3) removes any listed_taxa on the list for the project for taxa that 
@@ -21,10 +19,6 @@ The script:
    4) records the names of taxa in the EOL collection that aren't represented 
       by any iNat taxon_names (these taxa or taxon_names will be manually
       added)
-
-Improvements:
-  How would I automatically create the project_icon from the EOL logo_url?
-  More elegant way for stripping author of EOL Sci Name?
 
 Options:
 EOS
@@ -37,6 +31,7 @@ EOS
   opt :test, "Don't actually touch the db", :short => "-t", :type => :boolean
 end
 
+#unless eol_collection_url = "http://eol.org/api/collections/1.0/14791"
 unless eol_collection_url = ARGV[0]
   puts "You must specify the URL of an EOL collection of EOL collections"
   exit(0)
@@ -49,12 +44,28 @@ unless the_user
   exit(0)
 end
 
-#Get the Collections in the iNaturalist Collection from EOL
-unless response = Net::HTTP.get_response(URI.parse(eol_collection_url))
+eol_source = Source.find_by_in_text('EOL')
+eol_source ||= Source.create(
+  :in_text => 'EOL',
+  :citation => "Encyclopedia of Life. Available from http://www.eol.org.",
+  :url => "http://www.eol.org",
+  :title => 'Encyclopedia of Life'
+)
+
+# Get the Collections in the iNaturalist Collection from EOL
+response = Net::HTTP.get_response(URI.parse(eol_collection_url))
+unless response.is_a?(Net::HTTPOK)
   puts "couldn't access iNat Collection on EOL"
-  break
+  exit(0)
 end
+
 doc = Nokogiri::XML(response.body)
+
+if error_elt = doc.at('error')
+  puts "Error retrieving collection: #{error_elt.inner_text}"
+  exit(0)
+end
+
 eol_collection_ids = doc.xpath('//object_id[../object_type/text() = "Collection"]').map(&:text)
 eol_collection_ids.each do |eol_collection_id|
   # sleep(5)
@@ -62,10 +73,10 @@ eol_collection_ids.each do |eol_collection_id|
   taxa_removed = []         #in the project_list as a listed_taxa but no longer in the eol collection
   taxa_added = []    #in the eol collection so added to the the project_list as a listed_taxa
   taxa_missing = []  #in the eol collection but not added because it doesn't exist as a taxon in iNat (make manually)
+  url = "http://eol.org/api/collections/1.0/#{eol_collection_id}/?per_page=1000"
+  puts url
   begin
     #Get the collection
-    url = "http://eol.org/api/collections/1.0/#{eol_collection_id}/?per_page=1000"
-    puts url
     unless response = Net::HTTP.get_response(URI.parse(url))
       puts "\tcouldn't access Collection #{eol_collection_id} on EOL"
       next
@@ -105,7 +116,7 @@ eol_collection_ids.each do |eol_collection_id|
     project.title = project_name
     project.source_url = collection_url if project.source_url.blank?
     project.description = col.at('description').try(:text)
-    project.save
+    project.save unless opts[:test]
   else
     project = Project.new(
       :user_id => the_user.id, 
@@ -114,12 +125,14 @@ eol_collection_ids.each do |eol_collection_id|
       :description => col.at('description').try(:text),
       :project_type => "contest")
     project.project_observation_rules.build(:operator => "on_list?")
-    if logo_url
+    unless logo_url.blank?
       io = open(URI.parse(logo_url))
       project.icon = (io.base_uri.path.split('/').last.blank? ? nil : io)
     end
-    unless opts[:test]
-      if project.save
+    if opts[:test]
+      project.build_project_list
+    else
+      if !opts[:test] && project.save
         if opts[:header] || opts[:css]
           cp = project.build_custom_project
           if opts[:header] && tpl = (open(opts[:header]).read rescue nil)
@@ -141,72 +154,115 @@ eol_collection_ids.each do |eol_collection_id|
   #so we can later see if any of these listed_taxa aren't on the collection anymore
   #and must be removed
   the_list = project.project_list
-  listed_taxa_taxon_ids = the_list.listed_taxa.map{|lt| lt.taxon_id}
+  listed_taxa_taxon_ids = the_list.listed_taxa.all(:select => "id, taxon_id").map{|lt| lt.taxon_id}
   
   collection_item_dwc_names = []
   
-  col.xpath('//collection_items/item[object_type/text() = "TaxonConcept"]').each do |node|
-    list_item = TaxonName.strip_author(node.at('name').text)
-    collection_item_dwc_names << list_item
-    annotation = node.at('annotation').try(:text)
-    puts "\t name: #{list_item}"
-    
-    #Check to see if the list_item is a taxon_name associated with a taxon_id that already has a listed_taxon
-    existing = the_list.listed_taxa.first(:include => {:taxon => :taxon_names}, :conditions => [
-      "taxon_names.lexicon = ? AND taxon_names.name = ? AND listed_taxa.taxon_id IN (?)",
-      TaxonName::SCIENTIFIC_NAMES,
-      list_item.strip,
-      listed_taxa_taxon_ids
-    ])
-    if existing
-      puts "\t\t#{list_item} already on #{the_list}, updating..."
-      existing.update_attributes(:description => annotation)
-      listed_taxa_taxon_ids.delete(existing.taxon_id)
+  all_taxon_concepts = [col.xpath('//collection_items/item[object_type/text() = "TaxonConcept"]')]
+  proceed = true
+  i=2
+  while proceed
+    url = "http://eol.org/api/collections/1.0/#{eol_collection_id}/?per_page=1000&page=#{i}"
+    puts url
+    #Get the collection
+    unless response = Net::HTTP.get_response(URI.parse(url))
+      puts "\tcouldn't access Collection #{eol_collection_id} on EOL"
+      next
+    end
+    col2 = Nokogiri::XML(response.body)
+    taxon_concepts = col2.xpath('//collection_items/item[object_type/text() = "TaxonConcept"]')
+    if taxon_concepts.count == 0
+      proceed = false
     else
-      #find the taxon to make a listed taxon
-      taxon = Taxon.single_taxon_for_name(list_item)
-      taxon ||= Taxon.find_by_name(list_item.strip)
-      taxon = nil if taxon && taxon.taxon_names.detect{|tn| tn.name == list_item}.blank?
-      unless taxon
-        external_names = Ratatosk.find(list_item)
-        if match = external_names.detect{|en| en.name == list_item}
-          match.save unless opts[:test]
-          taxon = match.taxon
-          if taxon.valid? && !taxon.new_record?
-            puts "\t\tImported new taxon: #{taxon}"
-            taxon.send_later(:graft) unless opts[:test]
-          else
-            puts "\t\tFailed to import #{taxon}: #{taxon.errors.full_messages.to_sentence}"
+      all_taxon_concepts << taxon_concepts
+    end
+    i+=1
+  end
+  
+  all_taxon_concepts.each do |taxon_concepts|
+    taxon_concepts.each do |node|
+      list_item = TaxonName.strip_author(node.at('name').text)
+      object_id = node.at(:object_id).text
+      collection_item_dwc_names << list_item
+      annotation = node.at('annotation').try(:text)
+      puts "\t name: #{list_item}"
+      
+      # Check to see if the list_item is a taxon_name associated with a taxon_id that already has a listed_taxon
+      existing = the_list.listed_taxa.first(:include => {:taxon => :taxon_names}, :conditions => [
+        "taxon_names.lexicon = ? AND LOWER(taxon_names.name) = ?",
+        TaxonName::SCIENTIFIC_NAMES,
+        list_item.strip.downcase
+      ])
+      if existing
+        puts "\t\t#{list_item} already on #{the_list}, updating..."
+        existing.update_attributes(:description => annotation) unless opts[:test]
+        listed_taxa_taxon_ids.delete(existing.taxon_id)
+      else
+        #find the taxon to make a listed taxon
+        taxon = Taxon.single_taxon_for_name(list_item)
+        taxon ||= Taxon.find_by_name(list_item.strip)
+        taxon = nil if taxon && taxon.taxon_names.detect{|tn| tn.name == list_item}.blank?
+        unless taxon
+          begin
+            external_names = Ratatosk.find(list_item)
+          rescue Timeout::Error, StandardError, Exception => exc
+            puts "\t\tFailed to import #{taxon}: #{exc.message}"
             taxon = nil
           end
+          if external_names
+            if match = external_names.detect{|en| en.name == list_item}
+              match.save unless opts[:test]
+              taxon = match.taxon
+              if taxon.valid? && !taxon.new_record?
+                puts "\t\tImported new taxon: #{taxon}"
+                taxon.delay.graft_silently unless opts[:test]
+              else
+                puts "\t\tFailed to import #{taxon}: #{taxon.errors.full_messages.to_sentence}"
+                taxon = nil
+              end
+            end
+          end
         end
-      end
       
-      unless taxon
-        rank = case list_item.split.size
-        when 1 then "genus"
-        when 2 then "species"
-        else "subspecies"
+        unless taxon
+          rank = case list_item.split.size
+          when 1 then "genus"
+          when 2 then "species"
+          else "subspecies"
+          end
+          taxon = Taxon.new(
+            :name => list_item,
+            :rank => rank,
+            :creator_id => the_user.id,
+            :updater_id => the_user.id,
+            :source_url => "http://eol.org/pages/#{object_id}",
+            :source_identifier => object_id,
+            :source => eol_source
+          )
+          unless opts[:test]
+            if taxon.save
+              puts "\t\tCreated new taxon: #{taxon}"
+              taxon.delay.graft_silently
+            else
+              puts "\t\tFailed to create taxon: #{taxon.errors.full_messages.to_sentence}"
+              next
+            end
+          end
         end
-        taxon = Taxon.new(:name => list_item, :rank => rank)
-        taxon.save unless opts[:test]
-        puts "\t\tCreated new taxon: #{taxon}"
-        taxon.send_later(:graft) unless opts[:test]
-      end
-      
-      lt = ListedTaxon.new(:taxon_id => taxon.id, :list_id => the_list.id, :manually_added => true)
-      lt.save unless opts[:test]
-      #Record the taxon we just created a listed_taxon for under 'taxa_added'
-      listed_taxa_taxon_ids.reject!{ |taxon_id| taxon_id == taxon.id }
-      if lt.valid?
-        puts "\t\tCreated #{lt} for #{list_item}"
-        taxa_added << taxon.name
-      else
-        puts "\t\tFailed to create #{lt} for #{list_item}: #{lt.errors.full_messages.to_sentence}"
+        
+        lt = ListedTaxon.new(:taxon_id => taxon.id, :list_id => the_list.id, :manually_added => true)
+        lt.save unless opts[:test]
+        #Record the taxon we just created a listed_taxon for under 'taxa_added'
+        listed_taxa_taxon_ids.reject!{ |taxon_id| taxon_id == taxon.id }
+        if opts[:test] || lt.valid?
+          puts "\t\tCreated #{lt} for #{list_item}"
+          taxa_added << taxon.name
+        else
+          puts "\t\tFailed to create #{lt} for #{list_item}: #{lt.errors.full_messages.to_sentence}"
+        end
       end
     end
   end
-    
   # Anything left on listed_taxa_taxon_ids doesn't exist on the EOL 
   # collection so the iNat listed_taxon will be destroyed IFF there's a rule 
   # specifying that obs must be on the list
@@ -215,7 +271,7 @@ eol_collection_ids.each do |eol_collection_id|
       thelt = the_list.listed_taxa.first(:conditions => { :taxon_id => lt_taxon_id } )
       thelt.destroy unless opts[:test]
       # Record the taxon we just destroyed the listed_taxon for under 'taxa_removed'
-      taxa_removed << Taxon.find_by_id(lt_taxon_id).name
+      taxa_removed << (Taxon.find_by_id(lt_taxon_id).try(:name) || lt_taxon_id)
       puts "\tRemoved #{thelt} with taxon_id #{lt_taxon_id} which is no longer in the collection"
     end
   end
@@ -244,9 +300,12 @@ eol_collection_ids.each do |eol_collection_id|
   puts
 end
 
-Project.all(:conditions => "source_url LIKE 'http://eol.org/collections%'").each do |p|
-  unless p.source_url =~ /collections\/(#{eol_collection_ids.join('|')})$/
-    p.destroy unless opts[:test]
-    puts "Destroyed #{p}, no longer an EOL iNat collection"
+# Delete EOL projects that aren't in the response.  Skip if response was blank, which was probably erroneous.
+unless eol_collection_ids.blank?
+  Project.all(:conditions => "source_url LIKE 'http://eol.org/collections%'").each do |p|
+    unless p.source_url =~ /collections\/(#{eol_collection_ids.join('|')})$/
+      p.destroy unless opts[:test]
+      puts "Destroyed #{p}, no longer an EOL iNat collection"
+    end
   end
 end
